Move `ResolvedResourceDescriptionDns` to `shared` module
